1. add synchronization to OutputFile.cpp when dealing when performing writes. (the transfers can still happen asynchronosly)
2. Fix for server re-creating the output file everytime a write request was received. Now creates once and then opens in append mode as new data comes in

+++ src/server/ServeFile.cpp
@@ -247,6 +247,9 @@ ServeFile::ServeFile(std::string name, bool compress, uint64_t blkSize, uint64_t
 
         if (output) {
             std::experimental::filesystem::create_directories(std::experimental::filesystem::path(_name).parent_path());
+            std::ofstream file;
+            file.open(_name, std::fstream::binary);
+            file.close();
         }
         if (!output) {
             if (_blkSize > _size) {
@@ -297,9 +300,15 @@ ServeFile::~ServeFile() {
     _pool.terminate();
 
     if (_output && _remove) {
+        std::cout<<"removing: "<<_name<<std::endl;
         remove(_name.c_str());
     }
-    ((LocalFileCache*)_cache.getCacheByName(LOCALFILECACHENAME))->removeFile(_regFileIndex);
+    else{
+        std::cout<<" not removing: "<<_name<<std::endl;
+    }
+    if (!_output) {
+        ((LocalFileCache*)_cache.getCacheByName(LOCALFILECACHENAME))->removeFile(_regFileIndex);
+    }
     log(this) << _name << " closed" << std::endl;
 }
 
@@ -424,7 +433,7 @@ bool ServeFile::writeData(char *data, uint64_t size, uint64_t fp) {
             memcpy(odata, data, size);
             _pool.addTask([this, odata, size, fp] {
                 std::ofstream file;
-                file.open(_name, std::fstream::binary);
+                file.open(_name, std::fstream::binary|std::ofstream::app);
                 std::unique_lock<std::mutex> flock(_fileMutex);
                 file.seekp(fp, file.beg);
                 file.write(odata, size);
